qmt-gateway/pyproject.toml
[tool.poetry]
name = "qmt-gateway"
version = "0.1.0"
description = "QMT Gateway - 迅投QMT独立网关服务"
authors = ["aaron yang <aaron_yang@jieyu.ai>"]
readme = "README.md"
license = "MIT"
packages = [{ include = "qmt_gateway" }]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
python-fasthtml = "^0.12.36"
monsterui = "^1.0.36"
loguru = "^0.7.3"
apscheduler = "^3.10.4"
sqlite-utils = "^3.39"
bcrypt = "^5.0.0"
uvicorn = "^0.34.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.0"
black = "^22.3.0"
pytest = "^8.3.5"
pytest-cov = "^7.0.0"

[tool.poetry.scripts]
qmt-gateway = "qmt_gateway.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "D"]
